Animations in Main Level (#89)
Brought the door animations into main level and sorted them into proper folders, also created a separate art level for the artists
