+++ Assets/Animations.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bd9b548e06346e841a2cd7efe6bc3642
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
