+++ Assets/Animations/Doors.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 775ebbd9f39196b4b9d252b35186cbe1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
