+++ Assets/Animations/Doors/Door2Close.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 52f13e16ee6dbf34784888edf1af1cf0
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
