+++ Assets/Animations/Doors/Door2Control.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 57ae545c9f434c14aa06b497194deece
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
