+++ Assets/Animations/Doors/Door2Open.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 61ad1e1937ffd7a498fd2f84f32fb125
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
