+++ Assets/Animations/Doors/DoorController.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 22fb77e1fc87b5f43b7a6ba8b6c41726
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
