+++ Assets/Animations/Doors/DoorOpen.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a19dccf4c1efa3a41bcf5d74054b2bf2
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
