+++ Assets/Animations/Doors/NetworkDoor2.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 47b83d90ec47ac9489e370c90f207058
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
