+++ Assets/Animations/Doors/NetworkDoor2Close.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 057c484b4e0aeea48aaba1ef05eb9a2b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
