+++ Assets/Resources/NetworkedSceneObjects/DoorType_2.prefab
@@ -297,7 +297,7 @@ MonoBehaviour:
   m_SynchronizePosition: 1
   m_SynchronizeRotation: 1
   m_SynchronizeScale: 0
-  m_UseLocal: 1
+  m_UseLocal: 0
 --- !u!114 &6802119718052801610
 MonoBehaviour:
   m_ObjectHideFlags: 0
