+++ Assets/Scenes/Art Scenes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0c8e7f7f0849fc94da886141c2b76f41
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
