--- Assets/Scenes/DoorClose.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3b340080b753a9045ace20c64ad9d5c7
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
