+++ Assets/Scripts/GamePlaySystems/Doors.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 27dbe60b99734f146b8c8ec866ee5477
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
