+++ Assets/Scripts/GamePlaySystems/Doors/DoorMechanic.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: be150d69e819a9c48a99cb93da688377
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
