+++ Assets/Scripts/GamePlaySystems/Doors/DoorMechanic2.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bfac795fc345a78478bea13e9783460c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
