+++ ProjectSettings/EditorBuildSettings.asset
@@ -44,7 +44,7 @@ EditorBuildSettings:
   - enabled: 0
     path: Assets/Scenes/ProgrammerScenes/Owen/PhysicsPlayground.unity
     guid: 7c7d8d04d85772445a0f42beb0c1c4ae
-  - enabled: 0
+  - enabled: 1
     path: Assets/Scenes/Levels/MainLevel.unity
     guid: 923a5e8f201dddd4c8b72abc219ed812
   - enabled: 0
